+++ snow-spec/src/main.ts
@@ -1,9 +1,9 @@
 import fs from "node:fs/promises";
 import Path from "node:path";
+//import util from "node:util";
+import { analyze } from "./analyze";
 import { emit } from "./emit";
 import { parse } from "./parse";
-import { analyze } from "./analyze";
-//import util from "node:util";
 
 async function main() {
   const input = await fs.readFile(Path.resolve("debug.txt"), { encoding: "utf-8" });
@@ -11,7 +11,7 @@ async function main() {
   //console.log(util.inspect(syntaxNode, { depth: 30 }));
   const symbolNode = analyze(syntaxNode);
   const json = emit(symbolNode);
-  //console.log(json);
+  console.log(json);
 }
 main()
 .catch(err => { console.error(err); });
